Add create amendment example
