+++ decidim-core/lib/decidim/core/test.rb
@@ -27,3 +27,4 @@ require "decidim/core/test/shared_examples/searchable_results_examples"
 require "decidim/core/test/shared_examples/has_space_in_mcell_examples"
 require "decidim/core/test/shared_examples/railtie_examples"
 require "decidim/core/test/shared_examples/edit_link_shared_examples"
+require "decidim/core/test/shared_examples/amendable/amendment_creation_examples"
